fase_2/tarefa24_ia_embarcatech/pyproject.toml
[project]
name = "tarefa24-ia-embarcatech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pip>=25.3",
    "scikit-learn>=1.7.2",
    "split-folders>=0.5.1",
    "tensorflow>=2.20.0",
]
